Add the `geometry_to_mesh` calcfunction
This calcfunction uses `gmsh` to convert a geometry file into a mesh.
The dimension of the mesh can be specified through the second argument
of the calcfunction.

The `gmsh` binary and Python interface are installed through the `gmsh`
package through PyPI. Even though there is a Python interface, it is not
actually used, because it is currently not clear if and how the API
allows to initialize the geometry from file, instead of through explicit
commands. That is why for now the CLI is called through `os.subprocess`.
Ideally, the Python API can be used directly in the future.

Note that running `gmsh` on Ubuntu requires the `libglu1-mesa` system
package to be installed. This is added to the README and the CI workflow
installs it so that the tests can run.