Add the `convert_mesh` calcfunction
This calcfunction uses the `meshio` Python package to convert a mesh
file from one format to another.

The tests for the XDMF format could not compare the converted content
with a reference file because the XML output changes based on the Python
version that is used and the binary H5 file changes in content, most
likely because it contains the filepath of the associated XML file which
is a randomly generated one by the calcfunction.