Add the `GeometryPoisson` work chain
The purpose of this work chain is to solve the Poisson equation for a
geometry that is defined in a `.geo` file. For now the work chain only
implements the first two steps. It first computes the mesh from the
geometry using `gmsh`. The mesh file is then converted from the `.msh`
format to `.xdmf`, using the `meshio` library.